Change native.request to RawMessage
"The Native Object defines the native advertising opportunity available
for bid via this bid request. It will be included as a JSON-encoded
string in the bid request’s imp.native field or as a direct JSON object,
depending on the choice of the exchange. While OpenRTB 2.3/2.4 supports
only JSON-encoded strings, many exchanges have implemented a formal
object."